+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add Fabric API provisioning support
+- Add Fabric mods provisioning support
+
+### Changed
+- Update MDH to Cobbler
 
 ## 1.0.0 - 2025-11-29
 ### Added